refactor(core-effects): adopt stable effectid registry and update network
Refactor the core effect and registry model around a stable EffectId:\n\n- Introduce EffectId definitions and generated IDs in config/effect_ids.h.\n- Replace index-based effect storage with an explicit registry of\n  {EffectId, name, effect, legacyAdapter} entries in RendererActor.\n- Update CoreEffects, PatternRegistry and plugin metadata to register\n  and query by EffectId, with an explicit expected count.\n- Push EffectId through Commands, SystemState, CommandSerializer and\n  StateSerializer so sync and presets are keyed by stable IDs.\n- Update WebServer cached state, HTTP/WS codecs and request handlers to\n  expose EffectId-based effect IDs and zone state over the network.\n- Wire EncoderManager to use display_order helpers for front-panel\n  effect selection.\n- Integrate the new audio fields and validation metrics into the core\n  render loop and tests.\n\nThis establishes EffectId as the canonical identity for effects across\nfirmware, sync and network APIs, decoupling metadata from storage order.

+++ firmware/v2/src/effects/CoreEffects.h
@@ -112,14 +112,14 @@ void effectPulse(RenderContext& ctx);
  * @param renderer Pointer to RendererActor
  * @return Number of effects registered
  */
-uint8_t registerCoreEffects(RendererActor* renderer);
+uint16_t registerCoreEffects(RendererActor* renderer);
 
 /**
  * @brief Register ALL effects (core + LGP) with RendererActor
  * @param renderer Pointer to RendererActor
  * @return Total number of effects registered
  */
-uint8_t registerAllEffects(RendererActor* renderer);
+uint16_t registerAllEffects(RendererActor* renderer);
 
 } // namespace effects
 } // namespace lightwaveos
